+++ spec/lib/appsignal_spec.rb
@@ -132,6 +132,10 @@ describe Appsignal do
           Appsignal.tag_request.should be_false
         end
       end
+
+      it "should also listen to tag_job" do
+        Appsignal.should respond_to(:tag_job)
+      end
     end
 
     describe ".transactions" do
@@ -252,13 +256,26 @@ describe Appsignal do
     end
 
     describe ".send_exception" do
+      before { Appsignal::Pipe.stub(:current => false) }
+      let(:tags) { nil }
+
       it "should send the exception to AppSignal" do
-        agent = double
+        agent = double(:shutdown => true)
         Appsignal.stub(:agent).and_return(agent)
         agent.should_receive(:send_queue)
         agent.should_receive(:enqueue).with(kind_of(Appsignal::Transaction))
 
         Appsignal::Transaction.should_receive(:create).and_call_original
+      end
+
+      context "with tags" do
+        let(:tags) { {:a => 'a', :b => 'b'} }
+
+        it "should tag the request before sending" do
+          transaction = Appsignal::Transaction.create(SecureRandom.uuid, {})
+          Appsignal::Transaction.stub(:create => transaction)
+          transaction.should_receive(:set_tags).with(tags)
+        end
       end
 
       it "should not send the exception if it's in the ignored list" do
@@ -270,7 +287,7 @@ describe Appsignal do
         begin
           raise "I am an exception"
         rescue Exception => e
-          Appsignal.send_exception(e)
+          Appsignal.send_exception(e, tags)
         end
       end
     end